refactor directory layout for ignite and deprecated kafka
